src/FasterAI/worker/requirements.txt
# API
flask
flask-socketio
pika
redis
requests
python-socketio
simple-websocket
websocket-client